str::Builder/wstr::Builder: take int capHint instead of size_t
The capacity hint is a small count, so int is the natural type and avoids
size_t/int juggling at call sites. Cast size_t call-site values to int and
simplify str::ReplaceTemp's capHint computation (all-int) accordingly.

prompt: change Builder(size_t capHint) => Builder(int capHint); simplify casting in the capHint computation

+++ src/MobiDoc.cpp
@@ -859,7 +859,7 @@ bool MobiDoc::LoadForPdbReader(PdbReader* pdbReader) {
     }
 
     ReportIf(doc != nullptr);
-    doc = new str::Builder(docUncompressedSize);
+    doc = new str::Builder((int)docUncompressedSize);
     size_t nFailed = 0;
     for (size_t i = 1; i <= docRecCount; i++) {
         if (!LoadDocRecordIntoBuffer(i, *doc)) {
